pyproject.toml
[project]
name = "siad"
version = "0.1.0"
description = "Satellite Imagery Anomaly Detection - Multi-modal world model for infrastructure change detection"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
    { name = "SIAD Team" }
]

dependencies = [
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "earthengine-api>=0.1.400",
    "google-cloud-storage>=2.14.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "rasterio>=1.3.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "h5py>=3.10.0",
    "matplotlib>=3.8.0",
    "Pillow>=10.0.0",
    "tqdm>=4.66.0",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
]

[project.scripts]
siad = "siad.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--strict-markers --tb=short"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
